benchmark/requirements.txt
datasets>=2.14.0
soundfile>=0.12.0
librosa>=0.10.0
jiwer>=3.0.0
tqdm>=4.65.0
nemo_toolkit[asr]>=2.0.0
funasr>=1.3.0
peft>=0.7.0
faster-whisper>=1.0.0
sentencepiece>=0.2.0
qwen-asr>=0.1.0
